// src/schemas/invoiceSchema.ts
// schemas/invoiceSchema.ts
import { z } from 'zod';

export const invoiceSchema = z.object({
  invoiceNo: z.string().min(1, "Invoice number is required"),
  invoiceDetails: z.record(z.unknown()).refine(val => Object.keys(val).length > 0, {
    message: "Invoice details are required"
  }),
  customerId: z.number().min(1, "Customer ID is required"),
  generatedById: z.number().min(1, "Generated By ID is required"),
  status: z.string().min(1, "Invoice status is required"),  // Status can be 'Pending', 'Paid', 'Overdue', etc.
});

export type InvoiceFormData = z.infer<typeof invoiceSchema>;